+++ src/EventTypeDefs.ts
@@ -1,7 +1,7 @@
 import type { ErasingEventContext } from './brushes/Eraser';
 import type { Control } from './controls/control.class';
 import type { Point } from './point.class';
-import type { FabricObject } from './shapes/fabricObject.class';
+import type { FabricObject } from './shapes/Object/FabricObject';
 import type { Group } from './shapes/group.class';
 import type { IText } from './shapes/itext.class';
 import type { StaticCanvas } from './static_canvas.class';
@@ -9,9 +9,14 @@ import type { TOriginX, TOriginY, TRadian } from './typedefs';
 import type { saveObjectTransform } from './util/misc/objectTransforms';
 import type { Canvas } from './__types__';
 
-export type ModifierKey = 'altKey' | 'shiftKey' | 'ctrlKey';
+export type ModifierKey = keyof Pick<
+  MouseEvent | PointerEvent | TouchEvent,
+  'altKey' | 'shiftKey' | 'ctrlKey' | 'metaKey'
+>;
 
-export type TPointerEvent = MouseEvent | TouchEvent;
+export type TOptionalModifierKey = ModifierKey | null | undefined;
+
+export type TPointerEvent = MouseEvent | TouchEvent | PointerEvent;
 
 export type TransformAction<T extends Transform = Transform, R = void> = (
   eventData: TPointerEvent,
@@ -39,7 +44,7 @@ export type Transform = {
   target: FabricObject;
   action: string;
   actionHandler: TransformActionHandler;
-  corner: string;
+  corner: string | 0;
   scaleX: number;
   scaleY: number;
   skewX: number;
@@ -57,11 +62,18 @@ export type Transform = {
   height: number;
   shiftKey: boolean;
   altKey: boolean;
-  original: ReturnType<typeof saveObjectTransform>;
+  original: ReturnType<typeof saveObjectTransform> & {
+    originX: TOriginX;
+    originY: TOriginY;
+  };
 };
 
 export type TEvent<E extends Event = TPointerEvent> = {
   e: E;
+};
+
+type TEventWithTarget<E extends Event = TPointerEvent> = TEvent<E> & {
+  target: FabricObject;
 };
 
 export type BasicTransformEvent<E extends Event = TPointerEvent> = TEvent<E> & {
@@ -95,10 +107,10 @@ export type TransformEvent<T extends Event = TPointerEvent> =
     absolutePointer: Point;
   };
 
-type SimpleEventHandler<T extends Event = TPointerEvent> = TEvent<T> & {
-  target: FabricObject;
-  subTargets: FabricObject[];
-};
+type SimpleEventHandler<T extends Event = TPointerEvent> =
+  TEventWithTarget<T> & {
+    subTargets: FabricObject[];
+  };
 
 type InEvent = {
   previousTarget?: FabricObject;
@@ -108,8 +120,7 @@ type OutEvent = {
   nextTarget?: FabricObject;
 };
 
-type DragEventData = TEvent<DragEvent> & {
-  target: FabricObject;
+type DragEventData = TEventWithTarget<DragEvent> & {
   subTargets?: FabricObject[];
   dragSource?: FabricObject;
   canDrop?: boolean;
@@ -119,7 +130,7 @@ type DragEventData = TEvent<DragEvent> & {
 type DropEventData = DragEventData & { pointer: Point };
 
 type DnDEvents = {
-  dragstart: TEvent<DragEvent> & { target: FabricObject };
+  dragstart: TEventWithTarget<DragEvent>;
   drag: DragEventData;
   dragover: DragEventData;
   dragenter: DragEventData & InEvent;
@@ -178,12 +189,10 @@ export type ObjectEvents = ObjectPointerEvents &
   DnDEvents &
   ObjectModifiedEvents & {
     // selection
-    selected: {
-      e: TEvent;
+    selected: Partial<TEvent> & {
       target: FabricObject;
     };
-    deselected: {
-      e?: TEvent;
+    deselected: Partial<TEvent> & {
       target: FabricObject;
     };